third_party/blink/web_tests/http/tests/security/mixedContent/websocket/insecure-websocket-in-secure-page-worker.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script>

if (window.testRunner) {
    testRunner.setPopupBlockingEnabled(false);
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", false);
}

var test = async_test("Test that an insecure WebSocket connection is blocked when created in a worker created from HTTPS resource.");
test.step(function ()
{
    let openedWindow;
    window.addEventListener("message", test.step_func(function (evt)
    {
        assert_equals(evt.data, "DONE", "evt.data");
        openedWindow.close();
        test.done();
    }), false);

    openedWindow = window.open("https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-throw-on-construction-worker.html");
});

</script>
</head>
<body>
</body>
</html>
