third_party/WebKit/LayoutTests/http/tests/security/mixedContent/websocket/insecure-websocket-in-secure-page-worker-expected.txt

CONSOLE ERROR: [blocked] The page at 'https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-throw-on-construction-worker.html' was loaded over HTTPS, but ran insecure content from 'ws://127.0.0.1:8880/echo': this content should also be loaded over HTTPS.

This is a testharness.js-based test.
PASS Test that an insecure WebSocket connection is blocked when created in a worker created from HTTPS resource. 
Harness: the test ran to completion.
